==> site parsing/3-site_information_preprocessing.ipynb <==
import pandas as pd
import numpy as np

Load data from utf-8 decoding and except bad decoded data

site_data = pd.read_csv('data/sites.csv', header=None)
site_data.columns = ['position', 'url', 'title', 'description', 'keywords']
site_data

site_data = site_data[~site_data['title'].str.contains('[�]+', na=False)]
site_data

Load data from cp1251 decoding. There are 8 sites with other decoding I fill bad decoded data as NaN

site_recoded = pd.read_csv('data/sites-cp1251.csv', header=None)
site_recoded.columns = ['position', 'url', 'title', 'description', 'keywords']
site_recoded

site_data = pd.concat([site_data, site_recoded])
del(site_recoded)
site_data

na_sites = ['mrom-iman.ru', 'originalnie-podarki.com', 'encar.com', 'koronaaltaya.ru', 'my.jjwxc.net', 'himasoku.com', 'mimizun.com', 'mimizun.com']
for na_site in na_sites:  
    site_data.loc[site_data['url'] == na_site, 'title'] = np.NaN  
    site_data.loc[site_data['url'] == na_site, 'keywords'] = np.NaN
    site_data.loc[site_data['url'] == na_site, 'description'] = np.NaN

I miss sites that contains symbol � (urrecognized cyrillic symbol) in description and keywords, but not contain in title, but there are not to mach this sites and I fill description and keywords as NaN 

site_data.loc[site_data['description'].str.contains('[�]+', na=False), 'description'] = np.NaN
site_data.loc[site_data['description'].str.contains('[�]+', na=False), 'keywords'] = np.NaN
site_data.loc[site_data['keywords'].str.contains('[�]+', na=False), 'description'] = np.NaN
site_data.loc[site_data['keywords'].str.contains('[�]+', na=False), 'keywords'] = np.NaN

I put sites with empty cells into another DataFrame

non_sites = site_data[(site_data['title'].isna() & site_data['description'].isna() & site_data['keywords'].isna())]
site_data = site_data[~(site_data['title'].isna() & site_data['description'].isna() & site_data['keywords'].isna())]
[non_sites.shape, site_data.shape]

I put sites that return error into DataFrame with empty cells

normal = ['000', '043', '064', '100', '101', '102', '105', '108', '111', '112', '123', '125', '133', '149', '152', '161', '170', 
          '180', '220', '246', '250','301', '303', '314', '339', '360', '356', '365', '381', '409', '484', '500px', 
          '500 miles',  '555', '557', '578', '777', '986', '990', '911']
non_sites = pd.concat([non_sites, site_data[site_data['title'].str.match('\d{3}\D*$', na=False)]])
site_data = site_data[~site_data['title'].str.match('\d{3}\D*$', na=False)]
for norm in normal:    
    site_data = pd.concat([site_data, non_sites[non_sites['title'].str.contains(norm, na=False)]])
    non_sites = non_sites[~non_sites['title'].str.contains(norm, na=False)]
[non_sites.shape, site_data.shape]

normal = ['Строительный', 'Компания', 'Продажа', 'БридерШоп']
non_sites = pd.concat([non_sites, site_data[site_data['title'].str.contains('401', na=False)]])
site_data = site_data[~site_data['title'].str.contains('401', na=False)]
for norm in normal:    
    site_data = pd.concat([site_data, non_sites[non_sites['title'].str.contains(norm, na=False)]])
    non_sites = non_sites[~non_sites['title'].str.contains(norm, na=False)]
[non_sites.shape, site_data.shape]

I chose sites that contains text not in russian and english language

other_lang = site_data[site_data['title'].str.match('[^ A-Za-zА-Яа-я0-9_|%!—@*\'¡•"\∀)\(\]\[↓.:🧨🔥❤—🥇™👱‍♀️🤑🚩🚀🛒💗📟🚘►≡📚🙏✰☭#Є🎥🎓ᐅ🏛✈✪®【】▷🎀♫$☃🍀©❶❷❸¤+=ᐉ⋆🎮☆><~\}\{🍴😃💊🌟🤠😎✔ᐈ🛠🎬🎣♛☀️﻿🎅ЙЁёй🎉▲👷🔔🧁☕🔥🌏✐⚓🔝🦗🏠❌˙·•●♥~💫👇🏍⚙️📝☎☢★✿ܓ🚗📢🎁📌⭐👩‍⚕️👨‍👨‍👶‍👶🏆𝐖𝐄𝐋𝐂𝐎𝐌𝐄 𝐆𝐀𝐌𝐄𝐑†°•;☞/🏥💪🌎ⓁⒹ📡↪↩”“—|🏆😘🍎📙👨‍💻🙅‍♀️🍕🍔🍟❗️📍✌🌞🌴🌊ᚠᛇᛟᚷ🐠▶🐉🌾🏅💕➤👀💰☎🐰🏅📰💎🦷💥🎶👍🔶👉😊♥❀💕🎎💒🎲⛷️⚽️☀❂∑♦\\?Ꙭ🔧🚚🍲🥣📺µღ№⭐✅⚡️🎰«»➣\t\r\n\-]+', na=False)]
other_lang

sites = ['pribor-perm.ru', 'veira.net', 'dostoevskylib.ru', 'mtorrent.ru', 'studyinchinas.com', 'prostroymat.ru', 'gkhnsc.ru', 
  'kompotique.ru', 'farm-college.ru', 'howdyho.net', 'veira-souz.ru', 'burvin.by', 'veira.su', 'ha-ha.mirtesen.ru', 
  'kadashnikov.narod.ru', 'maksfera.clan.su', 'm-zadornov.ru', 'snowbd.ru', 'parkourperception.com', 'tadikars.ru', 
  'bovisklub.narod.ru', 'rotor-vlg.ru', 'glavceramica.ru', 'paraplan.directoria.biz', 'sokrov.diary.ru']
for site in sites:
  other_lang = other_lang[other_lang['url'] != site]
other_lang.shape

For this sites I set language as other 

other_lang['language'] = 'other'
site_data = site_data.merge(other_lang[['url', 'language']], how = 'left', on=['url'])
site_data

For sites that not contains cyrillic letters I set language as english 

site_data.loc[site_data['language'].isna() & (~site_data['title'].str.contains('[а-яА-Я]{2,}', na=False)), 'language'] = 'english'
site_data

For other sites I set language as russian

site_data['language'] = site_data['language'].fillna('russian')
site_data

I fill NA as empty lines

site_data = site_data.fillna('')
site_data

non_sites['title'] = ''
non_sites['description'] = ''
non_sites['keywords'] = ''
non_sites

site_data = pd.concat([site_data, non_sites])
site_data = site_data.fillna('unknown')
site_data

site_data.to_csv('data/site_information.csv', index=False)
