==> site parsing/4-text_preprocessing.ipynb <==
import sys
import os
import warnings
os.environ['OPENBLAS_NUM_THREADS'] = '1'
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
#for text pre-processing
import re, string
import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import SnowballStemmer
from nltk.corpus import wordnet
from nltk.stem import WordNetLemmatizer
!pip install pymorphy2
import pymorphy2
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
nltk.download('wordnet')
nltk.download('averaged_perceptron_tagger_ru')
#for model-building
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report, f1_score, accuracy_score, confusion_matrix
from sklearn.metrics import roc_curve, auc, roc_auc_score
# bag of words
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer
#for word embedding
import gensim
from gensim.models import Word2Vec

import pandas as pd
import numpy as np
import time
import pyarrow as pa
import pyarrow.parquet as pq
import scipy
!pip install implicit
import implicit
import bisect
import sklearn.metrics as m
!pip install catboost
from catboost import CatBoostClassifier, CatBoostRegressor, Pool
from sklearn.calibration import calibration_curve, CalibratedClassifierCV

import seaborn as sns
import matplotlib.pyplot as plt
import plotly.express as px

%matplotlib inline
sns.set_style('darkgrid')

!pip install feather-format >> none
!pip install faiss-cpu --no-cache

site_data = pd.read_csv('/data/site_information.csv')
site_data

site_data = site_data.rename(columns={'url' : 'url_host'})
site_data

site_data[['title', 'description', 'keywords']] = site_data[['title', 'description', 'keywords']].fillna('')
site_data

site_data = site_data.dropna()
site_data

site_data = site_data[site_data['language'] != 'unknown']
site_data

# STOPWORD REMOVAL
def stopword(string):
    a= [i for i in string.split() if i not in (stopwords.words('english') + stopwords.words('russian'))]
    return ' '.join(a)
#LEMMATIZATION
# Initialize the lemmatizer
wl = WordNetLemmatizer()
morph = pymorphy2.MorphAnalyzer()
 
# This is a helper function to map NTLK position tags
def get_wordnet_pos(tag):
    if tag.startswith('J'):
        return wordnet.ADJ
    elif tag.startswith('V'):
        return wordnet.VERB
    elif tag.startswith('N'):
        return wordnet.NOUN
    elif tag.startswith('R'):
        return wordnet.ADV
    else:
        return wordnet.NOUN
# Tokenize the sentence
def lemmatizer(string):    
    word_pos_tags = nltk.pos_tag(word_tokenize(string), lang='rus') # Get position tags
    list_of_words = []
    for tag in word_pos_tags:
        if tag[1] != 'NONLEX':
            list_of_words.append(morph.normal_forms(tag[0])[0])
        else:
            list_of_words.append(tag[0])
    word_pos_tags = nltk.pos_tag(list_of_words) # Get position tags
    a=[wl.lemmatize(tag[0], get_wordnet_pos(tag[1])) for tag in word_pos_tags] # Map the position tag and lemmatize the word/token
    return " ".join(a)

%%time
def finalpreprocess(string):
    return lemmatizer(stopword(string))
site_data[['title', 'description', 'keywords']] = site_data[['title', 'description', 'keywords']].applymap(lambda x: finalpreprocess(x))
site_data.head()

site_data['info'] = site_data['title'] + ' ' + site_data['description'] + ' ' + site_data['keywords']
site_data['info']

site_data['title_description'] = site_data['title'] + ' ' + site_data['description']
site_data['title_description']

site_data['title_keywords'] = site_data['title']+ ' ' + site_data['keywords']
site_data['title_keywords']

site_data['description_keywords'] = site_data['description'] + ' ' + site_data['keywords']
site_data['description_keywords']

site_data.to_csv('data/site_data.csv', index=False)
