==> site parsing/5-tf-idf_embendings.ipynb <==
import sys
import os
import warnings
os.environ['OPENBLAS_NUM_THREADS'] = '1'
warnings.filterwarnings('ignore')

import pandas as pd
import numpy as np
#for text pre-processing
import re, string
import nltk
from nltk.tokenize import word_tokenize
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import SnowballStemmer
from nltk.corpus import wordnet
from nltk.stem import WordNetLemmatizer
!pip install pymorphy2
import pymorphy2
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
nltk.download('wordnet')
nltk.download('averaged_perceptron_tagger_ru')
#for model-building
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report, f1_score, accuracy_score, confusion_matrix
from sklearn.metrics import roc_curve, auc, roc_auc_score
# bag of words
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer

import pandas as pd
import numpy as np
import time
import pyarrow as pa
import pyarrow.parquet as pq
import scipy
!pip install implicit
import implicit
import bisect
import sklearn.metrics as m
!pip install catboost
from catboost import CatBoostClassifier, CatBoostRegressor, Pool
from sklearn.calibration import calibration_curve, CalibratedClassifierCV

import seaborn as sns
import matplotlib.pyplot as plt
import plotly.express as px

%matplotlib inline
sns.set_style('darkgrid')

!pip install feather-format >> none
!pip install faiss-cpu --no-cache

site_data = pd.read_csv('/data/site_data.csv')

tfidf_vectorizer = TfidfVectorizer(use_idf=True)
title_tfidf = tfidf_vectorizer.fit_transform(site_data['title'].fillna(''))
title_tfidf

als = implicit.approximate_als.FaissAlternatingLeastSquares(factors = 50, iterations = 30, use_gpu = False, \
       calculate_training_loss = False, regularization = 0.1)

%%time
als.fit(title_tfidf)

u_factors = als.user_factors 
d_factors = als.item_factors
[len(u_factors), len(d_factors)]

%%time
url_dict = {url: idurl for url, idurl in zip(site_data['url_host'], range(len(site_data['url_host'])))}
inv_url_map = {v: k for k, v in url_dict.items()}
title_emb = pd.DataFrame(d_factors)
title_emb['url_host'] = title_emb.index.map(inv_url_map)
title_emb.to_csv('title_emb.csv', index=False)

tfidf_vectorizer = TfidfVectorizer(use_idf=True)
description_tfidf = tfidf_vectorizer.fit_transform(site_data['description'].fillna(''))
description_tfidf

als = implicit.approximate_als.FaissAlternatingLeastSquares(factors = 50, iterations = 30, use_gpu = False, \
       calculate_training_loss = False, regularization = 0.1)

%%time
als.fit(description_tfidf)

u_factors = als.user_factors 
d_factors = als.item_factors
[len(u_factors), len(d_factors)]

%%time
description_emb = pd.DataFrame(d_factors)
description_emb['url_host'] = description_emb.index.map(inv_url_map)
description_emb.to_csv('description_emb.csv', index=False)

tfidf_vectorizer = TfidfVectorizer(use_idf=True)
keywords_tfidf = tfidf_vectorizer.fit_transform(site_data['keywords'].fillna(''))
keywords_tfidf

als = implicit.approximate_als.FaissAlternatingLeastSquares(factors = 50, iterations = 30, use_gpu = False, \
       calculate_training_loss = False, regularization = 0.1)

%%time
als.fit(keywords_tfidf)

u_factors = als.user_factors 
d_factors = als.item_factors
[len(u_factors), len(d_factors)]

%%time
keywords_emb = pd.DataFrame(d_factors)
keywords_emb['url_host'] = keywords_emb.index.map(inv_url_map)
keywords_emb.to_csv('keywords_emb.csv', index=False)

tfidf_vectorizer = TfidfVectorizer(use_idf=True)
info_tfidf = tfidf_vectorizer.fit_transform(site_data['info'].fillna(''))
info_tfidf

als = implicit.approximate_als.FaissAlternatingLeastSquares(factors = 50, iterations = 30, use_gpu = False, \
       calculate_training_loss = False, regularization = 0.1)

%%time
als.fit(info_tfidf)

u_factors = als.user_factors 
d_factors = als.item_factors
[len(u_factors), len(d_factors)]

%%time
info_emb = pd.DataFrame(d_factors)
info_emb['url_host'] = info_emb.index.map(inv_url_map)
info_emb.to_csv('info_emb.csv', index=False)

tfidf_vectorizer = TfidfVectorizer(use_idf=True)
title_description_tfidf = tfidf_vectorizer.fit_transform(site_data['title_description'].fillna(''))
title_description_tfidf

als = implicit.approximate_als.FaissAlternatingLeastSquares(factors = 50, iterations = 30, use_gpu = False, \
       calculate_training_loss = False, regularization = 0.1)

%%time
als.fit(title_description_tfidf)

u_factors = als.user_factors 
d_factors = als.item_factors
[len(u_factors), len(d_factors)]

%%time
title_description_emb = pd.DataFrame(d_factors)
title_description_emb['url_host'] = title_description_emb.index.map(inv_url_map)
title_description_emb.to_csv('title_description_emb.csv', index=False)

tfidf_vectorizer = TfidfVectorizer(use_idf=True)
title_keywords_tfidf = tfidf_vectorizer.fit_transform(site_data['title_keywords'].fillna(''))
title_keywords_tfidf

als = implicit.approximate_als.FaissAlternatingLeastSquares(factors = 50, iterations = 30, use_gpu = False, \
       calculate_training_loss = False, regularization = 0.1)

%%time
als.fit(title_keywords_tfidf)

u_factors = als.user_factors 
d_factors = als.item_factors
[len(u_factors), len(d_factors)]

%%time
title_keywords_emb = pd.DataFrame(d_factors)
title_keywords_emb['url_host'] = title_keywords_emb.index.map(inv_url_map)
title_keywords_emb.to_csv('title_keywords_emb.csv', index=False)

tfidf_vectorizer = TfidfVectorizer(use_idf=True)
description_keywords_tfidf = tfidf_vectorizer.fit_transform(site_data['description_keywords'].fillna(''))
description_keywords_tfidf

als = implicit.approximate_als.FaissAlternatingLeastSquares(factors = 50, iterations = 30, use_gpu = False, \
       calculate_training_loss = False, regularization = 0.1)

%%time
als.fit(description_keywords_tfidf)

u_factors = als.user_factors 
d_factors = als.item_factors
[len(u_factors), len(d_factors)]

%%time
description_keywords_emb = pd.DataFrame(d_factors)
description_keywords_emb['url_host'] = description_keywords_emb.index.map(inv_url_map)
description_keywords_emb.to_csv('description_keywords_emb.csv', index=False)
